Updated 'nice names' for the various profiles, added 64-bit Debugger profile support, improved univ build support
At some point PPC support was dropped in openMSX for universal builds.
However, openMSX-Builder does still offer universal builds that include
PPC due to an override.

The openMSX Debugger can now also be build for the x86_64 platform and
support for that is also added in this commit. You will need to add
:target_cpu to your Debugger profile in your ~/.openMSX-builder.yaml
config though. See this commit's diff for example profiles.

The 'nice names' for the profiles have also been updated to more
clearly show for which platform the builds are.

Mac OS X 10.6 Snow Leopard also changed the 'hdiutil: create failed
error 49168' error message that randomly  occurs, and support for this
was added as well.